rearrange files: move GraphInput class and Input.txt to utilities package
