+++ src/main/java/org/shortest_path_problem/omsu/utilities/Input.txt
